Move `opCodeCanBeAdjustedTo` to `getLongDisplacementMnemonic`
`opCodeCanBeAdjustedTo` is not a very good name for what it does. It
looks for long displacement versions of non-long displacement
mnemonics. We rename the `opCodeCanBeAdjustedTo` API to something more
sensible, `getLongDisplacementMnemonic` and we move the API out of the
Instruction class as it is a mnemonic query and it does not belong
there.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>