Deprecate TR_GPR64 and use is64BitReg instead - Part I
This is Part I of the change which will be done in lock step. Some APIs
need to change and introduce more strict rules, and other code needs to
be deprecated. In this part we mainly focus on eliminating the uses of
`TR_GPR64` in all non-RA related codegen code.

On 31-bit platforms the Z codegen currently uses the `TR_GPR64`
register kind to represent that a virtual register holds a 64-bit
value. This is orthogonal to 64-bit platforms which use the
`is64BitReg` query to determine the answer to the same question.

Because the use of 64-bit registers on 31-bit has been enabled by
default this orthogonality is a regular source of weird RA bugs in
which we have to handle these separate paths always. Sometimes we miss
handle either the 31-bit or the 64-bit "way of doing things" and we
introduce functionally incorrect behavior.

To avoid this in the future we fold away the `TR_GPR64` register kind
and consolidate everything to use the 64-bit platform's way of doing
things, i.e. using the `is64BitReg` query.

This involves quite intricate work all over the codebase, and
predominantly in the register allocator (RA). Because of the strong
coupling between the use of 64-bit registers on 31-bit and the use of
the `TR_GPR64` and `is64BitReg` queries we are left with no choice but
to fold away the `use64BitRegsOn32Bit` query as always returning
`true`.

That is, following this change there will be no way of going back as
disabling use of 64-bit registers on 31-bit will be functionally
incorrect.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>