Deprecate TR_GPR64 and use is64BitReg instead - Part II
This is Part II of the change which will be done in lock step. In this
part we eliminate the last remaining uses of `TR_GPR64` in the register
allocator itself.

There seems to be issues throughout the register allocator that only
act on the `is64BitReg` flag if HPR support is enabled. This is however
not true because we can have 64-bit virtual registers being assigned
without HPRs being assigned. In such a case the HPR real registers in
the register allocator still need to point to the 64-bit virtual
registers so that on 31-bit we know to evict such values on linkage
transitions.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>