Distinguish between HPR assignment and HPR tracking
Now that 64-bit registers are readily used on 31-bit we need to
distinguish between the assignment of HPR registers, i.e. spilling,
filling, assignment, etc. with tracking of HPR registers w.r.t. the
use of 64-bit virtual registers.

That is to say if HPR assignment support is disable either due to
hardware restrictions or a supplied user option the register
allocator should still use HPR registers internally as we need to
keep track of 64-bit registers properly so that on 31-bit if we need
to spill HPRs across a linkage dispatch we know which HPRs are
assigned (meaning they point to 64-bit virtual registers) and we can
spill them properly.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>