Deprecate changeRegisterKind
This API as indicated by the comment is lightly tested. In fact the
code generator should now have the ability to change the register kind
as this is a highly complex topic only the register allocator
understands. Deprecating this API to avoid shooting ourselves in the
foot by mistakenly using this API without understanding what it does.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>