Deprecate UpperHalfIsDead register flag and all its dependencies
The isUpperHalfDead query is not used anywhere in the compiler.
Furthermore with HPRs on Z the flag is never reset if an HPR is used,
which means that it's functionally incorrect. We see no uses for this
flag to be set on every single (GPR) register that gets allocated thus
we deprecate the flag here and all the code that follows.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>