Convert RuntimeAssumptionTable to mark with lazy sweep
The current RuntimeAssumptionTable design relies on eager clean-up
and the clean-up code is dispresed in the codebase. This change does
two things: 1) it changes the runtime assumption table to use a
self-contained clean-up mechanism defaulting to lazy cleanup and 2)
isolate the API to traverse the RAT to simplify caller implementaiton
and ensure the clean-up is done in a consistent manner.

To faciliate the lazy clean-up of the RAT the special bit in the
next assumption for the JITed body is modified to be always useable.
APIs are added to traverse assumptions both ignoring and respecting
assumptions marked for removal.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>