Add JitBuilder services for unsigned division and remainder
Previously, JitBuilder had the Div and Rem services to perform signed
division and remainder. However, it was lacking services to perform the
unsigned versions of these operations. To fill this gap, the UnsignedDiv
and UnsignedRem services have been added.

Note that the lack of an lurem opcode at this time has required the
implementation of a hacky solution to allow calculation of the unsigned
remainder of two 64-bit integers. Right now, this is being implemented
by calculating (lhs - ((lhs / rhs) * rhs)). This does return the correct
value, but is obviously suboptimal and should be replaced if such an
opcode is ever added.

Signed-off-by: Ben Thomas <ben@benthomas.ca>