Add -qnocsect for z/OS compiles
On zOS XLC linker can't handle files with same name at link time.
A workaround with pragma is needed. What this does is essentially
give a different name to the codesection (csect) for this file. So it
doesn't conflict with another file with the same name.

Alternatively we can specify -qnocsect to avoid such issues. There
does not seem ot be any obvious downside to this.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>