Add double map API for LINUX and balanced GC policy
Add test for balanced policy to test double map API

Lengthy arrays in balanced policy are stored using arraylets.
To use such arrays, JNI critical copies index by index
from each array let leaf to a temporary array so that the
application has a contiguous representation of the array
(required by JNI critical). This operation, however, is
expensive.

For each large array, the introduced API will double map the
arraylet leaves associated with that array. The API created a
contiguous block of memory of the size of the entire array data,
and assigns parts of it to each arraylet leaf with calls to mmap.
Note that the heap must be shared and have a file descriptor
associated with it. This file descriptor is used to reference the
address of each arraylet leaf and double map them to the
contiguous region of memory created. So instead of needing to
copy every single element of the array to a temporary array
and then copy back, all modifications made to the newly created
Contiguous block of memory will be propagated to the arraylet
leaves in the heap.

Signed-off-by: Igor Braga <higorb1@gmail.com>