Implement a default snippet print function when tracing
There is no reason to assert because of an unknown snippet type when
tracing. Instead if we end up calling the base class virtual
implementation we will print a sane default value indicating to the
user that there is indeed a snippet there, however no custom printing
logic has been implemented yet.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>