Simplify prologue and epilogue creation on Linux system linkage
With the introduction of four new APIs:

- `fillGPRsInEpilogue`
- `fillFPRsInEpilogue`
- `spillGPRsInPrologue`
- `spillFPRsInPrologue`

We can migrate code from the the prologue and epilogue creation into
the aforementioned APIs to simplify the implementation. There can be
more work done within these functions, for example taking advantage of
the GPR/FPR masks which will be handled in subsequent commits.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>