Allow for Entry Point Marker to be NULL for callouts
We may be dispatching a call to an XPLINK function from a non-XPLINK
function, in which case we may not have an Entry Point Marker for the
non-XPLINK body. We still however need the call descriptor to be
present as the XPLINK body we are calling will return past the NOP
descriptor.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>