Deprecate TR_S390ProcessorInfo
Moving forward we will be using the CPU class for processor detection.
As an initial step we deprecate the `TR_S390ProcessorInfo` class in the
code generator. We will be modifying all the API usage in a separate
commit as there is quite a bit of code to be changed.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>