Fix z10- paths of generateShiftAndKeepSelected(32|64)Bit APIs
- Fix generateShiftAndKeepSelected64Bit by properly zeroing bits
- Fix generateShiftAndKeepSelected32Bit by properly zeroing bits and
loading the initial value from the second operand

The original intent of these APIs seem to be shift left and select,
judging by the use of the RISBG type instructions. The z10- paths doing
only a single shift left and shift right will not suffice. There are
probably other subtleties here such as what if the shift amount is
greater negative, or greater than `aToBit`, etc. See #3814 for some
more details and what we need to do to fix this.

Issue: #3814
Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>