Allow use of _ISOC99_SOURCE and __IBMCPP_TR1__ on the command line
Some existing pieces of code attempt to define this macro which enables
us to use specific APIs on z/OS such as `snprintf`, however they are
not checking if the define already exists on the command line which
results in redefinition errors. Fix this by checking if the define
already exists.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>