Change BC to an RX-a instruction format
Technically this is an RX-b instruction however we currently do not
have a proper class for such instructions and we generate the `BC`
instruction as part of z/OS linkage in which we use a real register to
generate the mask. Assertions later on prevent us from doing this if
the `BC` instruction is properly marked as an RX-a instruction.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>