Fix calculation of XPLINK call descriptor offset
The call descriptor offset in the NOP following a call within an XPLINK
frame was not being correctly calculated. The patch location was
correct, however the offset encoded was from the patch location, which
is incorrect.

As per the documentation linked in the source code the value we wish to
encoding is a (positive) signed offset in doublewords from  the
doubleword at or preceding the return point (NOP) to the XPLINK call
descriptor for this signature.

This is not something we can encode with the existing relocations, and
due to the doubleword alignment that is required it is not something
that we should generalize either since we don't envision other uses for
such a strict API. Instead we create a private relocation in the XPLINK
linkage class and handle the relocation there.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>