Fix XPLINK call descriptor generation for callouts
The cursor currently needs to be specified to `genCallNOPAndDescriptor`
to ensure it is generated after the call instruction. Currently because
of the way the `TR::S390Imm2Instruction` constructor works it ends up
generating the call descriptor at the start of the instruction stream
if a `NULL` preceeding instruction is specified. This will be fixed in
outside the scope of this change.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>