Use __malloc31 on z/OS to allocate code cache related memory
We currently lack full trampoline support in OMR. To get JitBuilder
callouts to native functions working we must ensure we can encode the
distance from the JIT method body to the native with a `BRASL`
instruction which only has a signed 32-bit range.

The only way to currently achieve this is to allocate all code cache
memory below the bar. Once JitBuilder is hooked up to the port library
we should be able to safely use the port library APIs to accomplish
this. See #4719 for more details to references of how this is done in
OpenJ9.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>