Add more parameters to JitBuilder SystemLinkageTest
On Linux on Z for example we can fit 4 integral and 4 floating point
arguments into registers for linkage. This means that no arguments are
ever passed through the stack. We want to test this path (it actually
has a bug which is fixed in the next commit) so we double the number of
arguments to the test to ensure some arguments are passed on the stack.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>