Fix LIBPATH on Jenkins pipeline on z/OS
Currently due to a limitation of the z/OS CMake port we require to
specify an explicit `LIBPATH` so that the various tests which
dynamically load DLLs can find them. Once CMake is fixed to properly
generate the dynamic libraries in the correct directories setting of
the `LIBPATH` environment variable should be removed. See #4761.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>