Enhance Tril LinkageTest to be much more comprehensive
The current LinkageTest is geared towards running on x86. Now that we
have full support for calls on z/OS this test can be enhanced. The
current test has implicit assumptions on the number of arguments that
are passed on the stack for a particular calling convention.

To be more exhaustive we greatly expand the scope of the testing here
to the following:

- JIT to {jit|native} calls to functions of different number of
arguments ranging from 1 to 9 with the same argument type
- JIT to {jit|native}calls to functions of different number of
arguments ranging from 1 to 9 with mixed argument types

This ends up running 136 different tests (+ 4 for invalid linkage)
which covers most cases for passing arguments to and from JIT methods.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>