Use method symbol address to get the entry point in Tril and JitBuilder
Some platforms, such as AIX and z/OS, use metadata structures to
represent "pointer to a function" retrieved via the C `&` operator. On
AIX and z/OS these are called Function Descriptors. We need to
disambiguate between the "pointer to a function" and "address of entry
point of a function".

Currently `getCodeStart` API returns the latter, however this is not
what we want when we query for the result of a compilation in Tril and
JitBuilder. We want the former, i.e. the value of the "pointer to a
function" of the JIT compiled method.

We use the method symbol `getAddress` API here to query for such
values, and ensuring this value is set properly after binary encoding.
It is currently set to the result of `getCodeStart` by default, so no
change in value. If platforms have Function Descriptors it is up to
those linkages to update this value once the location of the Function
Descriptor is known.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>