Fix printing of instruction prefixes before binary encoding
If one attempts to print an instruction prefix before binary encoding
we need to ensure that the cursor value is valid, otherwise we need to
fall back to printing the generic string.

More specifically, if one attempts to print the contents of a snippet
before it is binary encoded, the cursor value may be `NULL` or a small
positive value, since we are incrementing the cursor as we print the
snippet contents. We do not want to try dereferencing such cursor
pointers since they are not valid and would lead to segmentation faults.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>