Disable -Winvalid-offsetof in OMROptions and FEBase.cpp
See #1662 for some further details. This will not be easy to solve
becalse it is not straightforward to covert `TR::JitConfig` to a POD
so we can use `offsetof`. This will hopefully be solved via the options
framework overhaul, so fixing this warning properly is not worth the
effort at the moment. Instead we locally disable this warning around
the problematic area.