Fix CCN5562
```
CCN5562 (W) No code is generated for the asm statement. The asm statement is ignored.
```

Assembly generation for C/C++ code is disabled by default unless the
`-qasm` flag is set. The above warning was telling us the inline
assembly code was not generating anything, which is definitely not
what we want. Enabling `-qasm` further exposes bugs with the syntax
for inline assembly on xlC which is why the changes in AtomicSupport.cpp
were needed.