Suspend convlit around inline assembly on z/OS
If we are building with `-Wc,"convlit(ISO8859-1)"` xlC option then we
want to avoid doing literal conversion of inline assembly statements
because HLASM expects them to be in EBCDIC. Failing to do this results
in compilation errors with HLASM as it tries to interpret ASCII
characters of the inline assembly string.

This pragma is used elsewhere in the codebase so it is not something
new.