Fix CCN8924
The warning complains about passing an argument of non-POD type through
the following helper defined by Google Test:

```
static char (&Helper(...))[2];  // NOLINT
```

xlC for some reason does not like passing the non-POD type through this
helper, although it appears to work correctly. This helper is used for
statically determining whether types are implicitly covertable so all
we care about is whether `sizeof` this helper is 1 or 2. Whether the
type is passed through elipsis or as a parameter does not matter in this
case and hence we can just silence the warning on xlC.

```
CCN8924 (W) Cannot pass an argument of non-POD class type "const ASTNodeArg" through ellipsis.
```